// src/lib/miden/db/transaction-types.ts
export type ITransactionIcon =
  | 'SEND'
  | 'RECEIVE'
  | 'DEPLOY'
  | 'REJECTED'
  | 'MINT'
  | 'CONVERT_PRIVATE'
  | 'CONVERT_PUBLIC'
  | 'CONVERT_PRIVATE_TOKEN'
  | 'CONVERT_PUBLIC_TOKEN'
  | 'STAKE'
  | 'DEFAULT';